pyproject.toml
[tool.poetry]
name = "fearless"
version = "0.1.0"
description = ""
authors = ["Afshan Ejaz <afshanejaz90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pytest-playwright = "^0.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
